path and change location
