+++ test/us02_test.py
@@ -1,0 +1,31 @@
+import sys
+import unittest
+import os
+sys.path.append(os.path.abspath('../test/seeds'))
+sys.path.append(os.path.abspath('../src'))
+import us02
+
+
+class Testuser02(unittest.TestCase):
+    #all marriages after birth
+
+    def test01(self):
+        s = ""
+        self.assertEquals(s, us02.us02("seeds/test1.ged"))
+
+    def test02(self):
+        s = ""
+        self.assertEquals(s, us02.us02("seeds/test2.ged"))
+
+    def test03(self):
+        s = ""
+        self.assertEquals(s, us02.us02("seeds/test3.ged"))
+    def test04(self):
+        s = ""
+        self.assertEquals(s, us02.us02("seeds/test4.ged"))
+    def test05(self):
+        s = ""
+        self.assertEquals(s, us02.us02("seeds/test5.ged"))
+
+if __name__ == '__main__':
+    unittest.main()
